ASC-157 Fix free ratio globals
This commit fixes a bug in the `test_hypervisor_free.py` test.
Previously, the ratios were not set to use the global ratios in the
`test_hypervisor_free` test. This commit adds the global declarations in
the context of that test and ensures that they have a seed value set in
global namespace.