(ASC-618) fix cinder failure on newton
Prior to this PR the test that creates volume using command like `openstack volume create --size 1 volume_name`, without using argument `--availability-zone` as it is not a mandatory argument in latest openstack doc.

Create volume test passed in Master(Queens) and Pike, but failed in newly added Newton CI. Looking at the doc for Newton, the argument `--availability-zone` is in the example of create a volume (https://docs.openstack.org/newton/user-guide/common/cli-manage-volumes.html)

This PR to add the above argument to the test.