ASC-572 Skip instance_per_network_per_hypervisor
This commit skips the instance_per_network_per_hypervisor.py test.
Currently, this test fails on newton and newton-rc installations due to
the `openstack image list` command not returning any Ubuntu images.

Further research is required to determine the root cause of this failure
and to fix it. This work has been captured in ASC-639.