ASC-636 Deploy key pair to nova on utility
This commit updates the role tasks to add the ssh key to nova using the
utility container. The previous venv deployed to the network host does
not add any value and is eliminated. Application of the tasks has also
been throttled to just the first matching host is the network_hosts
group.