ASC-241 Create gateway network in converge
This commit add the creation of a public gateway network to the
network_setup task set. This network is intended to mirror the one
created by `openstack-ansible-ops`. This is done with the intention that
this network will not interfere with the deployment of that network when
the `openstack-asnible-ops` installation takes place in other molecule
testing suites. Conversely, the tasks in this commit will not execute if
they find that this network already exists. They assume that the gateway
network has been properly deployed if it is found to exist.

This strategy has been adopted because ensuring the removal of an
external gateway network can be quite tricky, especially if ports have
been opened to facilitate the use of DHCP or default gateway devices
owned by internal OpenStack network entities.