ASC-257 Catch JSON ValueError
This commit adds `_get_expected_value` helper. This is a copy of the
current helper in pytest-rpc. However, it traps the ValueError in the
event that the OpenStack query returns invalid JSON.

This trap should be added to the production pytest-rpc helper. When it
is available there, this commit should be reverted.