ASC-241 Delete servers before network setup
This commit fixes a logical error in the ordering of the ansible
OpenStack setup operations. This change ensures that all servers are
deleted prior to the network setup. This frees the network from any IP
bindings prior to delete/create. The network setup then happens prior to
server creation ensuring that the test network is available to bind to
prior to server creation.