ASC-241 Fix server network assignment
This commit fixes an error in the network assignment for the test
server. Prior to this commit, it was bound to a PRIVATE_NET that is not
managed by this role and will not be set up in a production run. The
assignment is now correctly set to the `test_network` ansible variable.