ASC-297 Delete swift fully rebalanced test
This commit deletes the "swift full rebalanced" test.

This test is seems like a mitigation step to perform if the previous
balance test does not result in balance values of less than 1.0. It is not
practical to wait 'a few hours' when running automated tests to validate that
the balance has reached a 0.00 value across all devices.