ASC-241 Use openstack-ansible-ops os-service-setup
This commit adds the use of the openstack-service-setup playbook from
the openstack-ansible-ops repo to deploy core resources.

This addition removes the need for additional image setup. This also
removes the need to manage the gateway network.

This addition brings this repo into parity with the primary setup used
by other molecule test repos.