ASC-266 Delete socks proxy access test
This commit deletes the test stub for the "Verify Kibana and Horizon are
accessible without SSH tunneling" test.

The purpose of this validation is to ensure that a customer deployment
is accessible via the rackspace bastion socks proxy. Since CI
deployments do not occur on a private external network, such a
validation cannot be properly executed in the CI gating jobs. Therefore,
this test is being removed from the suite.