ASC-790 Remove external-ip in security group rule
This commit removes the `--external-ip` flag from the security group
rule create commands in the converge tasks. This flag is not available
in the newton version of the openstack cli command and was causing the
execution of this task to fail. Because the flag was being set to its
default value, not having it in newer openstack deploys will not have
any effect on its execution.