ASC-791 Add network agents logic for newton
This commit updates the `test_dhcp_agents.py:test_openvswitch` test to
accommodate for OpenStack versions newton and older. Older versions of
`openstack` cli do not have the `--network` flag for the `openstack
network agent list` command. So, for these versions the legacy `neutron`
cli is used to list the agents associated with each network.