ASC-791 Add local helper for get_osa_version
This commit implements a local version of the `get_osa_version` in order
to determine the expected version values to query for. The helper in
pytest-rpc currently derives these values based on the branch of the
repo in which the executed test resides. However, the molecule tests are
submodules of the rpc-openstack-system-tests repo when executed in CI.
This means that all of the molecule tests are in a detached HEAD state
and cannot determine the properly expected branch.

This implementation assumes that the environment variable for
RE_JOB_BRANCH has been set to the desired value. An exception will be
thrown if the this variable is unset.

The tests that utilize this functionality have been updated to use the
local method.