ASC-790 rpc_product_release via ENV if fact unavai
This commit updates the logic for setting the `rpc_product_release`
fact. If the fact is not available from a local fact file, then the
value is assumed to be set via the RPC_PRODUCT_RELEASE environment
variable.

This fail over is necessary, because the fact file is not deployed on
the newton and newton-rc branches of rpc-openstack.