ASC-790 neutron cli if rpc_product_release unset
This commit updates the conditionals for choosing whether to use the
`openstack` or `neutron` cli commands for adding a default gateway to
the test router. If the fact is set to an empty string, this means that
neither a local fact or environment variable was found for the value.
So, this change fails to the more conservative command choice.