(ASC-806) Add test_create_and_assign_floating_ip back to nova test
This test case has been unreliable due to the test env was not clean. Recently works on cleaning the test env. bring hope that the test run will be more reliable, so we are trying to bring it back in.

(cherry picked from commit bd5994ab69cf49845bbd8611fcf8191dd9aecece)