ASC-1014 Delete rpc version test
This commit deletes the rpc version test meant to verify that the
expected major version of the installed RPC-O version matched that
associated with the branch from which it was installed.  The release
process process does not generate 'rc' tags in the git tree for
releases queens and newer.  Therefor, this test cannot be executed on
development builds.