(RI-471) added more server status checks and increased searching time for SSH
Prior to this PR, port 22 on newly created server cannot be reached using automated test, but it can be when doing it manually, it is possible that automated test run too fast that the newly created server ssh is not ready.

This PR is to increase the trial time from 100 seconds to 5 minutes, also add two more status checks on the newly created server: "OS-EXT-STS:power_state" and "OS-EXT-STS:vm_state" to ensure the server is ready before proceeding to SSH verification.