ASC-1043 Use ansible fact for rpc_product_release
This commit changes the lookup mechanism for the rpc_product_release
value from an environment variable to an ansible fact. In CI, the
environment variable falls out of scope prior to the execution of the
test. Since the ansible fact is defined, it becomes a more reliable
source for this value.

(cherry picked from commit 9e7aa1bf3260d86e11d47528ef3bd94d47dc03fb)