RI-417 Remove xfail tags
The two tests that were previously failing due the lack of being able to
SSH into the SUT have triggered failures via the xfail tag. This means
that the tests unexpectedly passed indicating that the issue has been
resolved. This commit removes these tags so that the tests will resume
there normal behavior.