(maint) increase running time for maas setup and verify tasks
Prior to this PR, verify task was set to run up to 30 minutes, unfortunately, it does not have enough time to complete the task.
This PR to increase both setup and verify tasks, also put a debug 'date' before and after the tasks to see how long they actually ran.