ASC-1053 Create custom fact - rpc_product_release
This commit creates a custom ansible fact to store the result of the
query of the rpc_product_release value regardless of whether that comes
from a deployed OSA custom fact or via an environment variable. This
custom fact is then available to the tests after the converge step has
concluded. The fact exists in the following hierarchy:

["ansible_facts"]["ansible_local"]["system_tests"]["rpc_product_release"]