(ASC-1097) Fix ci failure of unable to attach volume
Prior to this PR, attaching volume to a server fails because the volume is in 'building' state
This PR to add a check to ensure the volume is available and the server is ACTIVE, retry the check 30 times to bring the total of waiting for the volume to be available to 3 minutes. The test will be skipped if either server or volume is not in ideal state.