ASC-1052 Fix remove router subnet
Pike is currently failing when attempting to delete the subnets on all
existing routers during the converge step.  This is caused by the
"router show" information for the router not containing the key
"interfaces_info", which causes the condition to prematurely fail due
to the undefined variable used in the template.

This commit updates the code to query for the ports associated with the
routers. The "subnet_id" values associated with these ports are used
rather than the "interfaces_info" value. The router ids and there
subnets are mapped together in order to simplify the variables used for
the removal of the routers and their underlying components.