ASC-1052 Fix remove router subnet
Pike is currently failing when attempting to delete the subnets on all
existing routers during the converge step.  This is caused by the
"router show" information for the router not containing the key
"interfaces_info", which causes the condition to prematurely fail due
to the undefined variable used in the template.

This commit updates the code to query for the ports associated with the
routers. The "subnet_id" values associated with these ports are used
rather than the "interfaces_info" value. The router ids and there
subnets are mapped together in order to simplify the variables used for
the removal of the routers and their underlying components.

The queries and data construction are intended to implement the following:

router_list: list of routers as strings. These are used as dictionary keys
    for the "routers" fact.

output (from openstack port list): dictionary containing "results" key with
    list of ansible results from port list for each router. Each list element
    contains a dictionary of the results for one router. The "stdout_lines"
    key contains a list of the port data for the router. This data is then run
    through a regex filter to create a list of subnet_id values for the router.

Example inputs:
```
router_list=['first_router_id', 'second_router_id']
output={
           "results": [
               {
                   "stdout_lines": [
                       "first_port_key mac_addr ip_address='192.168.74.1', subnet_id='subnet_1a' ACTIVE"
                   ]
               },
               {
                   "stdout_lines": [
                       "first_port_key mac_addr ip_address='192.168.74.1', subnet_id='subnet_2a' ACTIVE"
                   ]
               }
           ]
       }
```
Would result in the following value for the `routers` fact:
```
routers={
            "first_router_id": [ "subnet_1a" ],
            "second_router_id": [ "subnet_2a" ]
        }
```