(ASC-1140) Added checking server creating status
System tests failed many tests in which a server needs to be created. During the server creation, the server status was seen to change from "Build" to "ERROR". It is unsure that there is at least one server has been successfully created, including the test_server created in server_setup.yml

This PR to add a step of checking if the server created in server_setup.yml is successful.