ASC-1052 Use with_subelements for router data
This commit alters the data structure of the routers fact in order to
enable the use of the `with_subelements` looping filter.

Prior to this commit, the subnets were being concatenated into the
`openstack router remove subnet` command. However, that command only
supports one subnet at a time. In order to use the `with_subelements`
facility to dynamically create multiple iterations of calling `openstack
router remove`, the data structure for the routers fact needed to be
altered so that known keys could be used for the filter. The structure
is now a list of dictionaries with the router id value stored under an
"id" key and the subnet list being stored under a "subnets" id.

This results in the following structure
```
routers:
  - id: "first_router_id"
    subnets:
      - "subnet_1a"
      - "subnet_1b"
  - id: "second_router_id"
    subnets:
      - "subnet_2a"
      - "subnet_2b"
```