ASC-1052 Ignore error on subnet removal
In CI there have been occasions where the router reports a spurious
subnet that cannot be deleted because it is not actually associated with
the router.

This commit ignores errors when removing subnets in order to work around
this particular case. If a subnet that is truly associated with the
router fails to get deleted, then the play will subsequently fail when
the router is attempted to be deleted. Therefore, skipping the failures
in these specific cases will not allow the converge step to falsely pass.

(cherry picked from commit 3ba5159fc56797a8c929e9c99040721deaf02931)