ASC-1204 Fix test_get_nova_allocation_ratios
This commit fixes an error in connecting to the `nova_conductor` host.
The host is now determined by introspecting the intersection of hosts in
the `nova_conductor` group that are also in the `infraX-host_containers`
group for the infraX target host.

(cherry picked from commit 8f383a9bea042a9e08e0283cb0fc971fc1955ba4)