ASC-1029 Generate clouds.yml on localhost
This commit adds the generation of the `clouds.yml` file locally in
order to facilitate running the `os_*` ansible modules in the converge
tasks.  The values for the `clouds.yml` file are derived from the
`overcloudrc` file on the director.  These are used to populate the
template that is included in this commit.