ASC-1030 OSPize write_to_attached_storage
This commit replaces the usage of `utility_container` with `cli_host` to
execute `openstack` commands in `test_verify_users_and_tenants.py`.

This commit also adds all previously used pytest-rpc helpers to local
`pytest_rpc_helpers.py` file.  The production helpers are bound to the
assumptions of an RPC-O deployment and needed to be altered to enable
the execution of `openstack` cli commands outside of that context.