ASC-1030 OSPize create_bootable_volume
This commit replaces the usage of `utility_container` with `cli_host` to
execute `openstack` commands in `test_create_bootable_volume.py`.

This commit also adds the previously used pytest-rpc helpers to the
local `pytest_rpc_helpers.py` file.  The production helpers are bound to
the assumptions of an RPC-O deployment and needed to be altered to
enable the execution of `openstack` cli commands outside of that
context.