ASC-1030 OSPize hypervisor_free
This commit replaces the usage of `utility_container` with `cli_host` to
execute `openstack` commands in `test_hypervisor_free.py`.

The OSP inventory does not contain a group for `nova_conductor`. So the
defined ratio values cannot be looked up from the system using that
node.  In this commit the values have been hard-coded to a value of `2`
in order to allow the test to pass in the absence of this information.
The correct source for this information should be determined and the
test should be updated to use this source accordingly (ASC-1257).