(ASC-1268) change scope for variable 'ssh_attempt'
Prior to this PR, variable 'ssh_attempt' was assigned inside a try-except-else
block, and the block is inside a 'for' loop. It is my understanding that if
the variable is not 'global' then the assignment alwasy go into the innermost
scope.
The test passed in local mnaio but failed on CI, that could be the reason that
depending on the python versions, the current varriable assignment might or
might not be accepted when the variable is assigned deeply inside a block
which is inside a 'for' loop.

This PR is simply declaring the variable before enter the 'for' loop. If it
still fail the next round of CI, then 'global' must be used.