(ASC-1268) Increase time searching for SSH
Prior to this commit, the test failed in CI while searching for SSH.
The error msg is "assert 'SSH' in ''". This error is really similar to
bug RI-471 where CI did not way long enough for SSH to come up, the fix
for that bug was simply increasing the searching time from 100 seconds
to 5 minutes.

That might be the same issue at this time, 5 minutes might be still too
short.

This PR to increase the searching time to 10 min.