ASC-1340 Harden lookup of rpc_product_release
This commit adds a method to perform a recursive search for a given key
in a given dictionary.  This method is then used to find the value of
the `rpc_product_release` ansible fact without needing to know where
this key is located in the fact data structure.

This solution was pulled from [stack overflow](https://stackoverflow.com/a/29652561)