(ASC-1350) fix attaching volume error
CI fails to assert len(volume['attachments']) == 1 because the volume['attachment']
is empty. The failure does make any sense because server status was confirmed 'ACTIVE'
and volume status is 'available' in the previous steps of the test, it could be an
intermittent failure.

This PR to increase the checking time in case it failed the server and volume status
checks in the future.