ASC-1314 Refactor 'test_create_snapshot_from_instance'
Introduced new fixtures for creating server instances via the OpenStack API.

Tech debt:

The 'openstack_properties' fixture was also changed to detach it from the
Ansible variables file. This was necessary because the moment a fixture pulls
in the testinfra 'host' fixture it will attempt to run all tests against every
host regardless if the test uses the 'host' fixture or not.

I also duplicated information inside the 'openstack_properties' fixture because
I'm planning on changing the naming conventions of the dictionary when all tests
have been refactored to utilize the OpenStack API.