ASC-1311 Refactor 'test_create_and_assign_floating_ip'
**conftest**

Updated the 'create_server' factory in 'conftest' to allow for specifying
'auto_ip' upon server creation. Discovered a better way to validate the
properties after server creation. Added a beginning sleep because OpenStack
lies about waiting until something is actually done.

Updated 'expect_os_property' to provide more detailed information if a property
was not found on the given OpenStack object.

Added a 'ping' helper with retry support.

**test_create_and_assign_floating_ip**

Added a beginning clean-up step for removing unattached floating IPs because it
was the easiest way to guarantee that the floating IP pool is not exhausted.