ASC-1341 Increase Floating IP Timeout
Turns out that attaching a floating IP is very difficult for OpenStack.
I increased the timeout to a ridiculous 10 minutes and moved the code for
attachment below verification that the server is running. The reason I think
failures occur only on the first attempt at creating a server with a floating
IP is because OpenStack uses aggressive caching.