ASC-1396 Use geneve network type not vxlan
OSP 13 deploys are currently not supporting `vxlan` network types.  The
`geneve` type appear to be the current default and supposedly supports all
of the functionality of `vxlan`.

This commit modifies the network type used by the openstack-ansible-ops
mnaio network creation script by dynamically replacing the `vxlan` value
for the `network_type` variable with the value of `geneve`.