ASC-1260 Delete cinder lvm test
Testing the underlying lvm/lvs storage of a cinder volume is specific to
the default backend storage method (lvm) on an openstack deployment. OSP
deployments use Ceph as its storage backend so testing lvm/lvs is not
appropriate OSP.

OSP also appears to put security measures in place to prevent directly
accessing the ceph rbd volumes on the overcloud-cephstorage nodes.
Therefore, it is not practical to create an analogous test to directly
introspect the rbd volumes associated with a cinder volume.