ASC-1327 Refactor 'test_swift_stat'
Added a new Ansible task file for copying the OpenStack version file to the
deployment host and adding an INI section header. The version file is then
parsed by the 'openstack_properties' fixture for OpenStack version semantics.

Updated the test to utilize the fixture for discovering the version of
OpenStack along with styling and docstring updates.