ASC-1512 Use PRIVATE_NET for instance from vol
This commit changes the network used for the instance created by
bootable volume.  Attempts to create such an instance using the
GATEWAY_NET as previously set consistently fail.  Using the PRIVATE_NET
successfully allows the instance to be created.  Since the network
binding is not significant to this test, it has no impact on the subject
under test.