ASC-1263 Use overcloud controller as neutron agent
OSP does not have an ansible group for 'neutron_agent', so performing
this lookup does not make sense in OSP.  This commit uses the
`overcloud-contoller-0` as the canonical resource for looking up the
`netns` commands to interact with network namespaces.

In addition, the criteria for a testable network is set to include the
existence of a dhcp agent associated with the network.  Otherwise the
`netns` commands do not make sense for such a network deployment.