ASC-1263 Create server per availability zone
This commit removes the use of `compute service list` hypervisors for
iterating on the creation of instances.  The availability zones
associated with these hypervisors are the same and knowledge of the
hypervisors does not provide control over which hypervisor the compute
instance is created.  Therefore, using them is wasteful and misleading.